requirements.txt
openai-agents
python-dotenv
requests
beautifulsoup4
tavily-python
